Load hiera model in float32 to avoid mixed-dtype conv2d error

+++ hiera/pytorch/loader.py
@@ -8,6 +8,7 @@ Hiera model loader implementation
 from typing import Optional
 from dataclasses import dataclass
 
+import torch
 from transformers import AutoModelForPreTraining, AutoImageProcessor
 
 from ...config import (
@@ -71,7 +72,9 @@ class ModelLoader(ForgeModel):
     def load_model(self, *, dtype_override=None, **kwargs):
         model_name = self._variant_config.pretrained_model_name
 
-        model = AutoModelForPreTraining.from_pretrained(model_name, **kwargs)
+        model = AutoModelForPreTraining.from_pretrained(
+            model_name, torch_dtype=torch.float32, **kwargs
+        )
         model.eval()
 
         if dtype_override is not None:
